+++ app/src/app.js
@@ -1,7 +1,6 @@
 const Koa = require('koa');
 const logger = require('logger');
 const koaLogger = require('koa-logger');
-const config = require('config');
 const loader = require('loader');
 const convert = require('koa-convert');
 const { RWAPIMicroservice } = require('rw-api-microservice-node');
@@ -52,13 +51,9 @@ app.use(koaLogger());
 app.use(koaSimpleHealthCheck());
 
 app.use(RWAPIMicroservice.bootstrap({
-    name: config.get('service.name'),
-    info: require('../microservice/register.json'),
-    swagger: require('../microservice/public-swagger.json'),
     logger,
-    baseURL: process.env.CT_URL,
-    url: process.env.LOCAL_URL,
-    token: process.env.CT_TOKEN,
+    gatewayURL: process.env.GATEWAY_URL,
+    microserviceToken: process.env.MICROSERVICE_TOKEN,
     fastlyEnabled: process.env.FASTLY_ENABLED,
     fastlyServiceId: process.env.FASTLY_SERVICEID,
     fastlyAPIKey: process.env.FASTLY_APIKEY
@@ -66,16 +61,7 @@ app.use(RWAPIMicroservice.bootstrap({
 
 loader.loadRoutes(app);
 
-const server = app.listen(process.env.PORT, () => {
-    if (process.env.CT_REGISTER_MODE === 'auto') {
-        RWAPIMicroservice.register().then(() => {
-            logger.info('CT registration process started');
-        }, (error) => {
-            logger.error(error);
-            process.exit(1);
-        });
-    }
-});
+const server = app.listen(process.env.PORT);
 
 logger.info('Server started in ', process.env.PORT);